Move split function to util
